refactor: useQuery setup option now receives the options as the second argument
BREAKING CHANGE: The setup option in useQuery now receives the options
as the second argument instead of a context object with the query return
value and the options. This allows the setup function to have a more
predictable signature and makes it easier to type. Only `PiniaColada`
has this option, **it has been removed from useQuery**. Overall, the
option still needs more thinking and will probably change in the future
again.