refactor!: rename `keys` to `invalidateKeys` and move to plugin
BREAKING CHANGE: The `keys` option that automatically invalidate keys
has been renamed to `invalidateKeys` and moved to a plugin. This is in
practice not needed. It's just an opinionated convenience that can be
replaced by directly invalidating queries in the cache with the
`onSettled()` hook in `useMutation()`:

```ts
const { mutate } = useMutation({
  onSettled({ caches, vars: { id } }) {
    caches.invalidateQueries({ key: ['contacts-search'] })
    caches.invalidateQueries({ key: ['contacts', id] })
  },
  mutation: (contact) => patchContact(contact),
})
```