fix(warn): avoid repeated queries warn
Close #192

I would like to revisit this in the future because it's still possible
to create scenarios where the values used in some of the options are
attached to an unmounted component and are therefore not reactive
anymore. For the moment, the warning is too frustrating for little
benefit so the docs have been updated instead.