fix: make options parameter optional in types
The types inside vue define that `Plugin<T>` where `T` isn't an array
implies that the options _must_ be passed.

When you do specify a tuple/array, you're defining the parameters of the
function. So `Plugin<[T0, T1]>` means `use(fn, T0, T1)`.

We can use this with an optional tuple member to make the parameter
optional: `Plugin<[T?]>` (which is basically `(...params: [T?])`).