Revert "refactor: deprecate onMutate in favor of onBeforeMutate"
This reverts commit 306a4f333686fad041a0fea4ec1eb3116c2a5809.
I was wrong regarding the language usage. As it turns out in English,
`on<event>` implies _run this when the event is about to happen, or as
it happens_. Therefore, the current naming is correct and it makes more
sense. Also _before_ often implies _after_ but there isn't one here
because we already have `onSuccess` and `onError`.