fix(query)!: invalidate inactive queries too
Fix #287

BREAKING CHANGE: `queryCache.invalidateQueries()` now accepts a second
parameter to control whether to refetch or not active queries. It can be
set to `'all'` to force fetch all queries. This replaces the existing
behavior of passing `active: null` (**can be removed now**) which wasn't
actually working. **You shouldn't be negatively affected by this change as
it wasn't working previously**.