fix(types): avoid incorrect undefined from tagged keys
BREAKING CHANGE: The `EntryKeyTagged` type now has multiple type params
and an array is no longer assignable to it. This is necessary to
correctly infer the types for `TData` and `TDataInitial` and if you were
**manually** using `EntryKeyTagged`, you will either need to cast arrays
to it or use `EntryKey` instead. **In most cases this should not affect
you**. This makes types stricter and also disallows setting a query data
to `undefined`. If you were doing
`queryCache.setQueryData(key, undefined)`, use `queryCache.setEntryState` instead.