fix(query): preserve any external AbortError
BREAKING CHANGE: while technically a fix, this is a breaking change if
you were relying on an external signal that aborts within the `query`.
It used to not set the state by just being an AbortError DomException
and now it will become an error. This behavior is more correct as we
don't want to ignore aborted signals that are external. Each fetch
creates an AbortController and passes the signal to `query` which is
aborted (without a reason) if any other method of the store fetches
again. This is done to avoid using an _outdated_ request.