refactor(mutations)!: keep mutations key as passed (#440)
BREAKING CHANGE: Mutations id are now just an incremented number that starts at 1. There is no longer a `$n` appended to keys for mutations and mutations without a key do not have an artificial key anymore. This is because initially the cache map was a more complex type but with it being a simple Map, there is no point in complexifying the keys. As a result the `mutationCache.get()` now takes the id of the mutation while `getEntries()` work the same. The `exact` filter has also been removed as mutations are, by nature, meant to be called multiple times.